course landing page #72
this adds in the course landing page to display to users allowing them
to select the course that they are either instructing or attending as a
student. the previous `Home.tsx` has been moved to
`course/StudentView.tsx` as that is what the page represents.

additional work has been done to include some starter code for the
`TeachersView` of a course that will be displayed if they are an
instructor for the course they are viewing. the page will display a WIP
message as that is being done in a different commit that will need to be
updated to match this.

also added in some error handling code for the api server so that we
will have a default error response and also catch any errors not
intentially caught in other code. the `/courses/student/:id` has been
removed and the `/courses/:uid` has been updated to display the
information needed for the new `/:course_id` page.