course enrolling / auto_enroll & registrations #80
added 2 new fields to a course to allow specifying if it is currently
allowing enrollments and if it will auto enroll users upon request. a
new mapping table has also been added that will contain a list of new
registrations for a course. currently the admin page is not setup to
display the list of registrations for a course. one will be added but
for now this will only allow toggling the enrolling / auto_enroll
fields.