progress interface update #86
updatd the student view to properly show the progress that they have for
the course. the view more reflects what the mockup shows and also
includes some changes that would help with indicating what demos have
been completed.

included an additional component in the event that the course has no
modules assigned to it. it will display some text and a link to take the
user back to the home page.

the Accordion component was not going to be sufficient and the necessary
components and logic have been manually inserted into the page to make
it work properly.