logic fixes #86
a previous assumption I had about how a module completion should be
calculated was wrong and updated the UI according to the expected
behavior.
also adjusted some potential issues around how the progress is
controlled. it will now be based on a rolling progression vs just
checking the previous one.
also included a percent indicator for a demo if it contains more than
one exercise and was not fully completed.