student registration #81
this adds in the ability for a student to go to a course and auto enroll
or register for a course. the course must be allowing enrollment and for
the auto_enroll the course must also be allowing that. the styling of
the page is fairly straight forward and modeled after other parts of the
site. errors are handled for the registration portion and if the student
is auto enrolled then they will receive course information once
enrolled so they do not have to reload the page.