Read page refactor #94
after having a look at the codeblock, I decided to refactor the page to
also include the changes needed for the codeblock. in another part of
the repo someone included a package that is the monaco editor but
wrapped in the necessary react code. I swapped over to using that and
adjusted what was needed to get the same functionality. also included
@tanstack/react-query and @tanstack/react-form to help with handling
forms and requests/responses to the api server. made various changes to
markup to reduced elements and some of the styling that was previousely
there.