demo assets update #93
this removes the static assets that were in place previousely with
information that is attached to the module. which leads to the issue
that if a demo is loaded without knowing what module it came from then
it will not display any related demos or resources. consolidated the
"GET" network requests into a single one that will retrieve all
information for the demo. also updated the demo likes to be in
transactions but was not sure how to reduce the network calls for the
ORM and just switched to raw sql queries.