exercise ordering #105
this adds in the ability to specify the order of exercises for a demo.
previousely there was no order to specify and most of the test demos
only had 1 exercise attached so it was not a problem. if more than one
exercise is specified then there was no specified order in which they
would be presented. a new field has been added that will specify the
desired order and the interface has been updated to easily allow adding
/ re-ordering / removing of exercises for a demo. this also includes
updates to the interface to use various tanstack libraries that have
been added to the project. other additions also include
 - updated the seed-db to use a transaction that will prevent partial
   data from being submitted to the database.
 - the youtube select was not properly retrieving all the videos to be
   shown. it will now retrieve the list of all videos for the channel.

   note: I did run across quota errors when testing it out which will
   result in an error being displayed vs showing any data. if that
   happens the field will not be displayed which will prevent the value
   from being updated. if this happens more often then we will need to
   look into storing the list of videos available to reduce api hits or
   find another solution.
 - the demo page will now sort the demos available by language and then
   title of the demo.

some of the styling for the form will be off as I was using changes that
are not included in this PR but will be in a different one.

the attempt page will also throw errors in this change. another PR
includes changes for that page that will be reviewed once this is out.