exercise prompt #106
this adds in the exercise prompt to the attempt page to be displayed
above the code input. added in some packages to help with parsing and
creating the react from the markdown. some of the code needed to
properly style the elements is not the best but for now works. I covered
most of the cases for markdown to render but if something is missing
then the styling can be added.